src/api/pyproject.toml
[tool.poetry]
name = "bk-user-api"
version = "2.3.0"
description = "bk-user Api"
authors = ["IMBlues <bluesedenyu@gmail.com>"]

[tool.poetry.dependencies]
python = "3.6.14"
aenum = "2.2.1"
Django = "3.2.5"
django-celery-beat = "2.2.0"
celery = "4.4.7"
drf-yasg = "1.20.0"
dataclasses = "0.7"
django-mptt = "0.12.0"
jsonfield = "3.1.0"
gunicorn = "19.9.0"
ldap3 = "2.6.1"
PyMySQL = "1.0.2"
phonenumbers = "8.11.0"
whitenoise = "5.2.0"
redis = "3.2.0"
openpyxl = "^3.0.3"
django-redis = "^4.11.0"
python-json-logger = "^0.1.11"
djangorestframework = "3.12.4"
djangorestframework-jsonp = "^1.0.2"
django-modeltranslation = "0.16.2"
# vine 5.0.0 may broke celery 4.3.0, fixed manually
vine = "1.3.0"
curlify = "^2.2.1"
bk-iam = "1.1.10"
regex = "2020.7.14"
gevent = "1.1.2"
cryptography = "^3.0.0"
pydantic = "^1.8.2"
django-environ = "^0.4.5"

[tool.poetry.dev-dependencies]
ipython = "^7.15.0"
# black
black = "^21.7b0"
# isort
isort = "^5.9.2"
# flake8
pyproject-flake8 = "^0.0.1-alpha.2"
flake8-comprehensions = "^3.5.0"
# pytest
pytest = "^6.2.4"
pytest-django = "^3.9.0"
pytest-cov = "^2.8.1"
# mypy
mypy = "^v0.910"
types-requests = "^2.25.0"
types-dataclasses = "^0.1.5"
types-pytz = "^2021.1.0"
types-pyyaml = "^5.4.3"
types-pymysql = "^1.0.0"
types-redis = "^3.5.4"
types-toml = "^0.1.3"

[[tool.poetry.source]]
name = "tencent-mirrors"
url = "https://mirrors.tencent.com/pypi/simple/"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "bkuser_core.config.overlays.dev"
addopts = "--disable-pytest-warnings --reuse-db"